+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 3.8.6
+- Fix bug collision. [#511](https://github.com/RafaelBarbosatec/bonfire/issues/511)
+- Renamed `AutomaticRandomMovement` to `RandomMovement`
+- `RandomMovement` improvements! Now works in `PLatformEnemy`
+
 # 3.8.5
 - Fix bug in `JumperAnimation`